Feat (?) : Sprint 2 service to calculate the adjustment with type

+++ RRADJUSTMENT/src/main/java/com/scor/rr/repository/AdjustmentnodeRepository.java
@@ -3,5 +3,5 @@ package com.scor.rr.repository;
 import com.scor.rr.domain.AdjustmentNodeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdjustmentnodeRepository extends JpaRepository<AdjustmentNodeEntity, Long> {
+public interface AdjustmentnodeRepository extends JpaRepository<AdjustmentNodeEntity, Integer> {
 }
